feat(react): match lucide icon names + add bare-name aliases
Rename 6 icons to lucide's canonical names (login→log-in, logout→log-out, snow-flake→snowflake, badge-dollar→badge-dollar-sign, dashboard→layout-dashboard, headphones-off→headphone-off) and yen-yuan→japanese-yen. Old names kept as aliases (component value AND handle type) so existing imports don't break.

Export bare lucide-style names (BellRing) alongside BellRingIcon for every icon, so @animateicons/react/lucide is a drop-in for lucide-react. Restore git-merge-conflict & japanese-yen; remove stale registry JSON (incl. pre-existing hu-bell orphan).
