+++ package.json
@@ -33,45 +33,45 @@
 	"dependencies": {
 		"@mdx-js/loader": "^3.1.1",
 		"@mdx-js/react": "^3.1.1",
-		"@next/mdx": "^16.2.9",
-		"@radix-ui/react-separator": "^1.1.10",
-		"@radix-ui/react-tooltip": "^1.2.10",
+		"@next/mdx": "^16.2.12",
+		"@radix-ui/react-separator": "^1.1.15",
+		"@radix-ui/react-tooltip": "^1.2.16",
 		"@types/mdx": "^2.0.14",
 		"@vercel/analytics": "^1.6.1",
 		"babel-plugin-react-compiler": "^1.0.0",
 		"class-variance-authority": "^0.7.1",
 		"clsx": "^2.1.1",
 		"date-fns": "^4.4.0",
-		"fuse.js": "^7.4.2",
+		"fuse.js": "^7.5.0",
 		"lucide-react": "^0.564.0",
-		"motion": "^12.40.0",
-		"next": "^16.2.9",
-		"radix-ui": "^1.6.0",
-		"react": "^19.2.7",
-		"react-dom": "^19.2.7",
+		"motion": "^12.43.0",
+		"next": "^16.2.12",
+		"radix-ui": "^1.6.7",
+		"react": "^19.2.8",
+		"react-dom": "^19.2.8",
 		"shiki": "^3.23.0",
-		"swr": "^2.4.1",
+		"swr": "^2.4.2",
 		"tailwind-merge": "^3.6.0",
 		"use-debounce": "^10.1.1"
 	},
 	"devDependencies": {
 		"@changesets/changelog-github": "^0.5.2",
-		"@changesets/cli": "^2.31.0",
-		"@tailwindcss/postcss": "^4.3.1",
+		"@changesets/cli": "^2.31.1",
+		"@tailwindcss/postcss": "^4.3.3",
 		"@testing-library/jest-dom": "^6.9.1",
 		"@testing-library/react": "^16.3.2",
-		"@types/node": "^25.9.3",
-		"@types/react": "^19.2.17",
-		"@types/react-dom": "^19.2.3",
+		"@types/node": "^25.9.5",
+		"@types/react": "^19.2.18",
+		"@types/react-dom": "^19.2.4",
 		"@vitejs/plugin-react": "^5.2.0",
-		"baseline-browser-mapping": "^2.10.37",
+		"baseline-browser-mapping": "^2.11.8",
 		"jsdom": "^26.1.0",
-		"prettier": "^3.8.4",
+		"prettier": "^3.9.6",
 		"prettier-plugin-tailwindcss": "^0.7.4",
-		"tailwindcss": "^4.3.1",
-		"tsx": "^4.22.4",
+		"tailwindcss": "^4.3.3",
+		"tsx": "^4.23.1",
 		"tw-animate-css": "^1.4.0",
 		"typescript": "^5.9.3",
-		"vitest": "^3.2.6"
+		"vitest": "^3.2.7"
 	}
 }
