--- public/r/hu-bell.json
@@ -1,16 +1,0 @@
-{
-	"$schema": "https://ui.shadcn.com/schema/registry-item.json",
-	"name": "bell",
-	"type": "registry:ui",
-	"addGlobalCss": false,
-	"registryDependencies": [],
-	"dependencies": ["motion"],
-	"devDependencies": [],
-	"files": [
-		{
-			"path": "bell-icon.tsx",
-			"content": "\"use client\";\n\nimport type { HTMLMotionProps, Variants } from \"motion/react\";\nimport { motion, useAnimation, useReducedMotion } from \"motion/react\";\nimport { forwardRef, useCallback, useImperativeHandle, useRef } from \"react\";\nimport { cn } from \"@/lib/utils\";\n\nexport interface BellIconHandle {\n startAnimation: () => void;\n stopAnimation: () => void;\n}\n\ninterface BellIconProps extends HTMLMotionProps<\"div\"> {\n size?: number;\n duration?: number;\n isAnimated?: boolean;\n}\n\nconst BellIcon = forwardRef<BellIconHandle, BellIconProps>(\n (\n  {\n   onMouseEnter,\n   onMouseLeave,\n   className,\n   size = 24,\n   duration = 1,\n   isAnimated = true,\n   ...props\n  },\n  ref,\n ) => {\n  const controls = useAnimation();\n  const reduced = useReducedMotion();\n  const isControlled = useRef(false);\n\n  useImperativeHandle(ref, () => {\n   isControlled.current = true;\n   return {\n    startAnimation: () =>\n     reduced ? controls.start(\"normal\") : controls.start(\"animate\"),\n    stopAnimation: () => controls.start(\"normal\"),\n   };\n  });\n\n  const handleEnter = useCallback(\n   (e?: React.MouseEvent<HTMLDivElement>) => {\n    if (!isAnimated || reduced) return;\n    if (!isControlled.current) controls.start(\"animate\");\n    else onMouseEnter?.(e as any);\n   },\n   [controls, reduced, isAnimated, onMouseEnter],\n  );\n\n  const handleLeave = useCallback(\n   (e: React.MouseEvent<HTMLDivElement>) => {\n    if (!isControlled.current) controls.start(\"normal\");\n    else onMouseLeave?.(e as any);\n   },\n   [controls, onMouseLeave],\n  );\n\n  const bellVariants: Variants = {\n   normal: { rotate: 0 },\n   animate: {\n    rotate: [0, -18, 15, -10, 6, -3, 0],\n    transition: {\n     duration: 1.6 * duration,\n     ease: \"easeInOut\",\n    },\n   },\n  };\n\n  const clapperVariants: Variants = {\n   normal: { x: 0 },\n   animate: {\n    x: [0, -4, 4, -2, 2, 0],\n    transition: {\n     duration: 1.6 * duration,\n     ease: \"easeInOut\",\n    },\n   },\n  };\n\n  return (\n   <motion.div\n    className={cn(\"relative inline-flex\", className)}\n    onMouseEnter={handleEnter}\n    onMouseLeave={handleLeave}\n    {...props}\n   >\n    <motion.svg\n     xmlns=\"http://www.w3.org/2000/svg\"\n     width={size}\n     height={size}\n     viewBox=\"0 0 24 24\"\n     fill=\"none\"\n     stroke=\"currentColor\"\n     strokeWidth=\"2\"\n     strokeLinecap=\"round\"\n     strokeLinejoin=\"round\"\n     animate={controls}\n     initial=\"normal\"\n     variants={bellVariants}\n    >\n     <motion.path\n      d=\"M15.5 18C15.5 19.933 13.933 21.5 12 21.5C10.067 21.5 8.5 19.933 8.5 18\"\n      variants={clapperVariants}\n     />\n     <path d=\"M19.2311 18H4.76887C3.79195 18 3 17.208 3 16.2311C3 15.762 3.18636 15.3121 3.51809 14.9803L4.12132 14.3771C4.68393 13.8145 5 13.0514 5 12.2558V9.5C5 5.63401 8.13401 2.5 12 2.5C15.866 2.5 19 5.634 19 9.5V12.2558C19 13.0514 19.3161 13.8145 19.8787 14.3771L20.4819 14.9803C20.8136 15.3121 21 15.762 21 16.2311C21 17.208 20.208 18 19.2311 18Z\" />\n    </motion.svg>\n   </motion.div>\n  );\n },\n);\n\nBellIcon.displayName = \"BellIcon\";\nexport { BellIcon };\n",
-			"type": "registry:ui"
-		}
-	]
-}
