--- public/r/lu-dashboard.json
@@ -1,18 +1,0 @@
-{
-  "$schema": "https://ui.shadcn.com/schema/registry-item.json",
-  "name": "dashboard",
-  "type": "registry:ui",
-  "addGlobalCss": false,
-  "registryDependencies": [],
-  "dependencies": [
-    "motion"
-  ],
-  "devDependencies": [],
-  "files": [
-    {
-      "path": "dashboard-icon.tsx",
-      "content": "\"use client\";\n\nimport { cn } from \"@/lib/utils\";\nimport type { Variants } from \"motion/react\";\nimport {\n LazyMotion,\n domMin,\n m,\n useAnimation,\n useReducedMotion,\n} from \"motion/react\";\nimport {\n forwardRef,\n useCallback,\n useImperativeHandle,\n useRef,\n type HTMLAttributes,\n} from \"react\";\nexport interface DashboardIconHandle {\n startAnimation: () => void;\n stopAnimation: () => void;\n}\n\ninterface DashboardIconProps extends Omit<\n HTMLAttributes<HTMLDivElement>,\n | \"color\"\n | \"onDrag\"\n | \"onDragStart\"\n | \"onDragEnd\"\n | \"onAnimationStart\"\n | \"onAnimationEnd\"\n | \"onAnimationIteration\"\n> {\n size?: number;\n duration?: number;\n isAnimated?: boolean;\n color?: string;\n}\n\nconst DashboardIcon = forwardRef<DashboardIconHandle, DashboardIconProps>(\n (\n  {\n   onMouseEnter,\n   onMouseLeave,\n   className,\n   size = 24,\n   duration = 0.6,\n   isAnimated = true,\n   color,\n   ...props\n  },\n  ref,\n ) => {\n  const controls = useAnimation();\n  const reduced = useReducedMotion();\n  const isControlled = useRef(false);\n\n  useImperativeHandle(ref, () => {\n   isControlled.current = true;\n   return {\n    startAnimation: () =>\n     reduced ? controls.start(\"normal\") : controls.start(\"animate\"),\n    stopAnimation: () => controls.start(\"normal\"),\n   };\n  });\n\n  const handleEnter = useCallback(\n   (e?: React.MouseEvent<HTMLDivElement>) => {\n    if (!isAnimated || reduced) return;\n    if (!isControlled.current) controls.start(\"animate\");\n    else onMouseEnter?.(e as any);\n   },\n   [controls, reduced, isAnimated, onMouseEnter],\n  );\n\n  const handleLeave = useCallback(\n   (e?: React.MouseEvent<HTMLDivElement>) => {\n    if (!isControlled.current) controls.start(\"normal\");\n    else onMouseLeave?.(e as any);\n   },\n   [controls, onMouseLeave],\n  );\n\n  const smoothEase: [number, number, number, number] = [0.22, 1, 0.36, 1];\n\n  const iconVariants: Variants = {\n   normal: { scale: 1, rotate: 0 },\n   animate: {\n    scale: [1, 1.06, 0.98, 1],\n    rotate: [0, -1.5, 1.5, 0],\n    transition: { duration: 1.1 * duration, ease: \"easeInOut\" },\n   },\n  };\n\n  const tileVariants: Variants = {\n   normal: { opacity: 1, scale: 1, y: 0 },\n   animate: (i: number) => ({\n    opacity: [0.6, 1],\n    scale: [0.95, 1.04, 1],\n    y: [3, -2, 0],\n    transition: {\n     duration: 0.9 * duration,\n     ease: \"easeInOut\",\n     delay: i * 0.08,\n    },\n   }),\n  };\n\n  return (\n   <LazyMotion features={domMin} strict>\n    <m.div\n     className={cn(\"inline-flex items-center justify-center\", className)}\n     onMouseEnter={handleEnter}\n     onMouseLeave={handleLeave}\n     {...props}\n     style={{ color, ...props.style }}\n    >\n     <m.svg\n      xmlns=\"http://www.w3.org/2000/svg\"\n      width={size}\n      height={size}\n      viewBox=\"0 0 24 24\"\n      fill=\"none\"\n      stroke=\"currentColor\"\n      strokeWidth=\"2\"\n      strokeLinecap=\"round\"\n      strokeLinejoin=\"round\"\n      animate={controls}\n      initial=\"normal\"\n      variants={iconVariants}\n     >\n      <m.rect\n       width=\"7\"\n       height=\"9\"\n       x=\"3\"\n       y=\"3\"\n       rx=\"1\"\n       variants={tileVariants}\n       custom={0}\n       initial=\"normal\"\n       animate={controls}\n      />\n      <m.rect\n       width=\"7\"\n       height=\"5\"\n       x=\"14\"\n       y=\"3\"\n       rx=\"1\"\n       variants={tileVariants}\n       custom={1}\n       initial=\"normal\"\n       animate={controls}\n      />\n      <m.rect\n       width=\"7\"\n       height=\"9\"\n       x=\"14\"\n       y=\"12\"\n       rx=\"1\"\n       variants={tileVariants}\n       custom={2}\n       initial=\"normal\"\n       animate={controls}\n      />\n      <m.rect\n       width=\"7\"\n       height=\"5\"\n       x=\"3\"\n       y=\"16\"\n       rx=\"1\"\n       variants={tileVariants}\n       custom={3}\n       initial=\"normal\"\n       animate={controls}\n      />\n     </m.svg>\n    </m.div>\n   </LazyMotion>\n  );\n },\n);\n\nDashboardIcon.displayName = \"DashboardIcon\";\nexport { DashboardIcon };\n",
-      "type": "registry:ui"
-    }
-  ]
-}
