--- public/r/lu-headphones-off.json
@@ -1,18 +1,0 @@
-{
-  "$schema": "https://ui.shadcn.com/schema/registry-item.json",
-  "name": "headphones-off",
-  "type": "registry:ui",
-  "addGlobalCss": false,
-  "registryDependencies": [],
-  "dependencies": [
-    "motion"
-  ],
-  "devDependencies": [],
-  "files": [
-    {
-      "path": "headphones-off-icon.tsx",
-      "content": "\"use client\";\n\nimport { cn } from \"@/lib/utils\";\nimport type { Variants } from \"motion/react\";\nimport {\n LazyMotion,\n domMin,\n m,\n useAnimation,\n useReducedMotion,\n} from \"motion/react\";\nimport {\n forwardRef,\n useCallback,\n useImperativeHandle,\n useRef,\n type HTMLAttributes,\n} from \"react\";\nexport interface HeadphonesOffIconHandle {\n startAnimation: () => void;\n stopAnimation: () => void;\n}\n\ninterface HeadphonesOffIconProps extends Omit<\n HTMLAttributes<HTMLDivElement>,\n | \"color\"\n | \"onDrag\"\n | \"onDragStart\"\n | \"onDragEnd\"\n | \"onAnimationStart\"\n | \"onAnimationEnd\"\n | \"onAnimationIteration\"\n> {\n size?: number;\n duration?: number;\n isAnimated?: boolean;\n color?: string;\n}\n\nconst HeadphonesOffIcon = forwardRef<\n HeadphonesOffIconHandle,\n HeadphonesOffIconProps\n>(\n (\n  {\n   onMouseEnter,\n   onMouseLeave,\n   className,\n   size = 24,\n   duration = 1,\n   isAnimated = true,\n   color,\n   ...props\n  },\n  ref,\n ) => {\n  const controls = useAnimation();\n  const reduced = useReducedMotion();\n  const isControlled = useRef(false);\n\n  useImperativeHandle(ref, () => {\n   isControlled.current = true;\n   return {\n    startAnimation: () =>\n     reduced ? controls.start(\"normal\") : controls.start(\"animate\"),\n    stopAnimation: () => controls.start(\"normal\"),\n   };\n  });\n\n  const handleEnter = useCallback(\n   (e?: React.MouseEvent<HTMLDivElement>) => {\n    if (!isAnimated || reduced) return;\n    if (!isControlled.current) controls.start(\"animate\");\n    else onMouseEnter?.(e as any);\n   },\n   [controls, reduced, isAnimated, onMouseEnter],\n  );\n\n  const handleLeave = useCallback(\n   (e?: React.MouseEvent<HTMLDivElement>) => {\n    if (!isControlled.current) controls.start(\"normal\");\n    else onMouseLeave?.(e as any);\n   },\n   [controls, onMouseLeave],\n  );\n\n  const headphonesVariants: Variants = {\n   normal: { rotate: 0, scale: 1 },\n   animate: {\n    scale: [1, 1.1, 0.95, 1],\n    rotate: [0, -3, 3, -2, 0],\n    transition: { duration: 1.2 * duration, ease: \"easeInOut\", repeat: 0 },\n   },\n  };\n\n  const earcupVariants: Variants = {\n   normal: { scale: 1, opacity: 1 },\n   animate: {\n    scale: [1, 1.2, 0.9, 1],\n    opacity: [1, 0.7, 1],\n    transition: { duration: 0.9 * duration, ease: \"easeInOut\", repeat: 0 },\n   },\n  };\n\n  const slashVariants: Variants = {\n   normal: { pathLength: 1, opacity: 1 },\n   animate: {\n    pathLength: [1, 0, 1],\n    opacity: 1,\n    transition: { duration: 1.2 * duration, ease: \"easeInOut\" },\n   },\n  };\n\n  return (\n   <LazyMotion features={domMin} strict>\n    <m.div\n     className={cn(\"inline-flex items-center justify-center\", className)}\n     onMouseEnter={handleEnter}\n     onMouseLeave={handleLeave}\n     {...props}\n     style={{ color, ...props.style }}\n    >\n     <m.svg\n      xmlns=\"http://www.w3.org/2000/svg\"\n      width={size}\n      height={size}\n      viewBox=\"0 0 24 24\"\n      fill=\"none\"\n      stroke=\"currentColor\"\n      strokeWidth=\"2\"\n      strokeLinecap=\"round\"\n      strokeLinejoin=\"round\"\n      animate={controls}\n      initial=\"normal\"\n      variants={headphonesVariants}\n     >\n      <m.path\n       d=\"M3 14h3a2 2 0 0 1 2 2v3a2 2 0 0 1-2 \n\t            2H5a2 2 0 0 1-2-2v-7a9 9 0 0 1 18 \n\t            0v7a2 2 0 0 1-2 2h-1a2 \n\t            2 0 0 1-2-2v-3a2 2 0 0 1 \n\t            2-2h3\"\n       variants={earcupVariants}\n      />\n\n      <m.path d=\"M22 2L2 22\" variants={slashVariants} strokeWidth={2.5} />\n     </m.svg>\n    </m.div>\n   </LazyMotion>\n  );\n },\n);\n\nHeadphonesOffIcon.displayName = \"HeadphonesOffIcon\";\nexport { HeadphonesOffIcon };\n",
-      "type": "registry:ui"
-    }
-  ]
-}
