--- public/r/lu-logout.json
@@ -1,18 +1,0 @@
-{
-  "$schema": "https://ui.shadcn.com/schema/registry-item.json",
-  "name": "logout",
-  "type": "registry:ui",
-  "addGlobalCss": false,
-  "registryDependencies": [],
-  "dependencies": [
-    "motion"
-  ],
-  "devDependencies": [],
-  "files": [
-    {
-      "path": "logout-icon.tsx",
-      "content": "\"use client\";\n\nimport { cn } from \"@/lib/utils\";\nimport type { Variants } from \"motion/react\";\nimport {\n LazyMotion,\n domMin,\n m,\n useAnimation,\n useReducedMotion,\n} from \"motion/react\";\nimport {\n forwardRef,\n useCallback,\n useImperativeHandle,\n useRef,\n type HTMLAttributes,\n} from \"react\";\nexport interface LogoutIconHandle {\n startAnimation: () => void;\n stopAnimation: () => void;\n}\n\ninterface LogoutIconProps extends Omit<\n HTMLAttributes<HTMLDivElement>,\n | \"color\"\n | \"onDrag\"\n | \"onDragStart\"\n | \"onDragEnd\"\n | \"onAnimationStart\"\n | \"onAnimationEnd\"\n | \"onAnimationIteration\"\n> {\n size?: number;\n duration?: number;\n isAnimated?: boolean;\n color?: string;\n}\n\nconst LogoutIcon = forwardRef<LogoutIconHandle, LogoutIconProps>(\n (\n  {\n   onMouseEnter,\n   onMouseLeave,\n   className,\n   size = 28,\n   duration = 1,\n   isAnimated = true,\n   color,\n   ...props\n  },\n  ref,\n ) => {\n  const controls = useAnimation();\n  const reduced = useReducedMotion();\n  const isControlled = useRef(false);\n\n  useImperativeHandle(ref, () => {\n   isControlled.current = true;\n   return {\n    startAnimation: () =>\n     reduced ? controls.start(\"normal\") : controls.start(\"animate\"),\n    stopAnimation: () => controls.start(\"normal\"),\n   };\n  });\n\n  const handleEnter = useCallback(\n   (e?: React.MouseEvent<HTMLDivElement>) => {\n    if (!isAnimated || reduced) return;\n    if (!isControlled.current) controls.start(\"animate\");\n    else onMouseEnter?.(e as any);\n   },\n   [controls, reduced, isAnimated, onMouseEnter],\n  );\n\n  const handleLeave = useCallback(\n   (e?: React.MouseEvent<HTMLDivElement>) => {\n    if (!isControlled.current) controls.start(\"normal\");\n    else onMouseLeave?.(e as any);\n   },\n   [controls, onMouseLeave],\n  );\n\n  const iconVariants: Variants = {\n   normal: { scale: 1, rotate: 0 },\n   animate: {\n    scale: [1, 1.1, 0.95, 1],\n    rotate: [0, 3, -2, 0],\n    transition: {\n     duration: 0.9 * duration,\n     ease: \"easeInOut\",\n    },\n   },\n  };\n\n  const arrowVariants: Variants = {\n   normal: { x: 0, opacity: 1 },\n   animate: {\n    x: [8, -2, 0],\n    opacity: [0, 1, 1],\n    transition: {\n     duration: 0.6 * duration,\n     ease: \"easeOut\",\n    },\n   },\n  };\n\n  const doorVariants: Variants = {\n   normal: { pathLength: 1 },\n   animate: {\n    pathLength: [0, 1],\n    transition: {\n     duration: 0.7 * duration,\n     ease: \"easeInOut\",\n     delay: 0.1,\n    },\n   },\n  };\n\n  return (\n   <LazyMotion features={domMin} strict>\n    <m.div\n     className={cn(\"inline-flex items-center justify-center\", className)}\n     onMouseEnter={handleEnter}\n     onMouseLeave={handleLeave}\n     {...props}\n     style={{ color, ...props.style }}\n    >\n     <m.svg\n      xmlns=\"http://www.w3.org/2000/svg\"\n      width={size}\n      height={size}\n      viewBox=\"0 0 24 24\"\n      fill=\"none\"\n      stroke=\"currentColor\"\n      strokeWidth=\"2\"\n      strokeLinecap=\"round\"\n      strokeLinejoin=\"round\"\n      animate={controls}\n      initial=\"normal\"\n      variants={iconVariants}\n     >\n      <m.path d=\"m16 17 5-5-5-5\" variants={arrowVariants} />\n      <m.path d=\"M21 12H9\" variants={arrowVariants} />\n\n      <m.path\n       d=\"M9 21H5a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h4\"\n       variants={doorVariants}\n      />\n     </m.svg>\n    </m.div>\n   </LazyMotion>\n  );\n },\n);\n\nLogoutIcon.displayName = \"LogoutIcon\";\nexport { LogoutIcon };\n",
-      "type": "registry:ui"
-    }
-  ]
-}
