--- public/r/lu-snow-flake.json
@@ -1,18 +1,0 @@
-{
-  "$schema": "https://ui.shadcn.com/schema/registry-item.json",
-  "name": "snow-flake",
-  "type": "registry:ui",
-  "addGlobalCss": false,
-  "registryDependencies": [],
-  "dependencies": [
-    "motion"
-  ],
-  "devDependencies": [],
-  "files": [
-    {
-      "path": "snow-flake-icon.tsx",
-      "content": "\"use client\";\n\nimport { cn } from \"@/lib/utils\";\n\nimport {\n LazyMotion,\n domMin,\n m,\n useAnimation,\n useReducedMotion,\n} from \"motion/react\";\nimport {\n forwardRef,\n useCallback,\n useImperativeHandle,\n useRef,\n type HTMLAttributes,\n} from \"react\";\nexport interface SnowFlakeIconHandle {\n startAnimation: () => void;\n stopAnimation: () => void;\n}\n\ninterface SnowFlakeIconProps extends Omit<\n HTMLAttributes<HTMLDivElement>,\n | \"color\"\n | \"onDrag\"\n | \"onDragStart\"\n | \"onDragEnd\"\n | \"onAnimationStart\"\n | \"onAnimationEnd\"\n | \"onAnimationIteration\"\n> {\n size?: number;\n duration?: number;\n isAnimated?: boolean;\n color?: string;\n}\n\nconst SnowFlakeIcon = forwardRef<SnowFlakeIconHandle, SnowFlakeIconProps>(\n (\n  {\n   onMouseEnter,\n   onMouseLeave,\n   className,\n   size = 24,\n   duration = 1,\n   isAnimated = true,\n   color,\n   ...props\n  },\n  ref,\n ) => {\n  const controls = useAnimation();\n  const reduced = useReducedMotion();\n  const isControlled = useRef(false);\n\n  useImperativeHandle(ref, () => {\n   isControlled.current = true;\n   return {\n    startAnimation: () =>\n     reduced ? controls.start(\"normal\") : controls.start(\"animate\"),\n    stopAnimation: () => controls.start(\"normal\"),\n   };\n  });\n\n  const handleEnter = useCallback(\n   (e?: React.MouseEvent<HTMLDivElement>) => {\n    if (!isAnimated || reduced) return;\n    if (!isControlled.current) controls.start(\"animate\");\n    else onMouseEnter?.(e as any);\n   },\n   [controls, reduced, isAnimated, onMouseEnter],\n  );\n\n  const handleLeave = useCallback(\n   (e: React.MouseEvent<HTMLDivElement>) => {\n    if (!isControlled.current) {\n     controls.start(\"normal\");\n    } else {\n     onMouseLeave?.(e as any);\n    }\n   },\n   [controls, onMouseLeave],\n  );\n\n  const pathVariants = {\n   normal: {\n    pathLength: 1,\n    opacity: 1,\n    transition: { duration: 0.3 * duration },\n   },\n   animate: {\n    pathLength: [1, 0.3, 1],\n    opacity: [1, 0.7, 1],\n    transition: { duration: 0.8 * duration },\n   },\n  };\n\n  return (\n   <LazyMotion features={domMin} strict>\n    <m.div\n     className={cn(\"inline-flex items-center justify-center\", className)}\n     onMouseEnter={handleEnter}\n     onMouseLeave={handleLeave}\n     {...props}\n     style={{ color, ...props.style }}\n    >\n     <m.svg\n      xmlns=\"http://www.w3.org/2000/svg\"\n      width={size}\n      height={size}\n      viewBox=\"0 0 24 24\"\n      fill=\"none\"\n      stroke=\"currentColor\"\n      strokeWidth=\"2\"\n      strokeLinecap=\"round\"\n      strokeLinejoin=\"round\"\n      variants={{\n       normal: {\n        rotate: 0,\n        scale: 1,\n        transition: { duration: 0.3 * duration },\n       },\n       animate: {\n        rotate: [0, 10, -10, 0],\n        scale: [1, 1.05, 1],\n        transition: { duration: 1 * duration },\n       },\n      }}\n      animate={controls}\n      initial=\"normal\"\n     >\n      <m.path d=\"m10 20-1.25-2.5L6 18\" variants={pathVariants} />\n      <m.path d=\"M10 4 8.75 6.5 6 6\" variants={pathVariants} />\n      <m.path d=\"m14 20 1.25-2.5L18 18\" variants={pathVariants} />\n      <m.path d=\"m14 4 1.25 2.5L18 6\" variants={pathVariants} />\n      <m.path d=\"m17 21-3-6h-4\" variants={pathVariants} />\n      <m.path d=\"m17 3-3 6 1.5 3\" variants={pathVariants} />\n      <m.path d=\"M2 12h6.5L10 9\" variants={pathVariants} />\n      <m.path d=\"m20 10-1.5 2 1.5 2\" variants={pathVariants} />\n      <m.path d=\"M22 12h-6.5L14 15\" variants={pathVariants} />\n      <m.path d=\"m4 10 1.5 2L4 14\" variants={pathVariants} />\n      <m.path d=\"m7 21 3-6-1.5-3\" variants={pathVariants} />\n      <m.path d=\"m7 3 3 6h4\" variants={pathVariants} />\n     </m.svg>\n    </m.div>\n   </LazyMotion>\n  );\n },\n);\n\nSnowFlakeIcon.displayName = \"SnowFlakeIcon\";\nexport { SnowFlakeIcon };\n",
-      "type": "registry:ui"
-    }
-  ]
-}
