--- public/r/lu-yen-yuan.json
@@ -1,18 +1,0 @@
-{
-  "$schema": "https://ui.shadcn.com/schema/registry-item.json",
-  "name": "yen-yuan",
-  "type": "registry:ui",
-  "addGlobalCss": false,
-  "registryDependencies": [],
-  "dependencies": [
-    "motion"
-  ],
-  "devDependencies": [],
-  "files": [
-    {
-      "path": "yen-yuan-icon.tsx",
-      "content": "\"use client\";\n\nimport { cn } from \"@/lib/utils\";\nimport type { Variants } from \"motion/react\";\nimport {\n LazyMotion,\n domMin,\n m,\n useAnimation,\n useReducedMotion,\n} from \"motion/react\";\nimport {\n forwardRef,\n useCallback,\n useImperativeHandle,\n useRef,\n type HTMLAttributes,\n} from \"react\";\nexport interface YenYuanIconHandle {\n startAnimation: () => void;\n stopAnimation: () => void;\n}\n\ninterface YenYuanIconProps extends Omit<\n HTMLAttributes<HTMLDivElement>,\n | \"color\"\n | \"onDrag\"\n | \"onDragStart\"\n | \"onDragEnd\"\n | \"onAnimationStart\"\n | \"onAnimationEnd\"\n | \"onAnimationIteration\"\n> {\n size?: number;\n duration?: number;\n isAnimated?: boolean;\n color?: string;\n}\n\nconst YenYuanIcon = forwardRef<YenYuanIconHandle, YenYuanIconProps>(\n (\n  {\n   onMouseEnter,\n   onMouseLeave,\n   className,\n   size = 24,\n   duration = 1,\n   isAnimated = true,\n   color,\n   ...props\n  },\n  ref,\n ) => {\n  const controls = useAnimation();\n  const reduced = useReducedMotion();\n  const isControlled = useRef(false);\n\n  useImperativeHandle(ref, () => {\n   isControlled.current = true;\n   return {\n    startAnimation: () =>\n     reduced ? controls.start(\"normal\") : controls.start(\"animate\"),\n    stopAnimation: () => controls.start(\"normal\"),\n   };\n  });\n\n  const handleEnter = useCallback(\n   (e?: React.MouseEvent<HTMLDivElement>) => {\n    if (!isAnimated || reduced) return;\n    if (!isControlled.current) controls.start(\"animate\");\n    else onMouseEnter?.(e as any);\n   },\n   [controls, reduced, isAnimated, onMouseEnter],\n  );\n\n  const handleLeave = useCallback(\n   (e?: React.MouseEvent<HTMLDivElement>) => {\n    if (!isControlled.current) controls.start(\"normal\");\n    else onMouseLeave?.(e as any);\n   },\n   [controls, onMouseLeave],\n  );\n\n  const svgVariants: Variants = {\n   normal: { scale: 1, rotate: 0, y: 0 },\n   animate: {\n    scale: [1, 1.06, 1],\n    rotate: [0, -2, 2, 0],\n    y: [0, -1, 0],\n    transition: { duration: 1.2 * duration, ease: \"easeInOut\" },\n   },\n  };\n\n  const vStroke: Variants = {\n   normal: { pathLength: 1, opacity: 1 },\n   animate: {\n    pathLength: [0, 1],\n    opacity: [0.7, 1],\n    transition: {\n     duration: 0.9 * duration,\n     ease: \"easeInOut\",\n     delay: 0.06,\n    },\n   },\n  };\n\n  const midStroke: Variants = {\n   normal: { pathLength: 1, opacity: 1 },\n   animate: {\n    pathLength: [0, 1],\n    opacity: [0.8, 1],\n    transition: {\n     duration: 0.6 * duration,\n     ease: \"easeInOut\",\n     delay: 0.22,\n    },\n   },\n  };\n\n  const baseStroke: Variants = {\n   normal: { pathLength: 1, opacity: 1 },\n   animate: {\n    pathLength: [0, 1],\n    opacity: [0.8, 1],\n    transition: {\n     duration: 0.6 * duration,\n     ease: \"easeInOut\",\n     delay: 0.34,\n    },\n   },\n  };\n\n  return (\n   <LazyMotion features={domMin} strict>\n    <m.div\n     className={cn(\"inline-flex items-center justify-center\", className)}\n     onMouseEnter={handleEnter}\n     onMouseLeave={handleLeave}\n     {...props}\n     style={{ color, ...props.style }}\n    >\n     <m.svg\n      xmlns=\"http://www.w3.org/2000/svg\"\n      width={size}\n      height={size}\n      viewBox=\"0 0 24 24\"\n      fill=\"none\"\n      stroke=\"currentColor\"\n      strokeWidth=\"2\"\n      strokeLinecap=\"round\"\n      strokeLinejoin=\"round\"\n      animate={controls}\n      initial=\"normal\"\n      variants={svgVariants}\n      className=\"lucide lucide-japanese-yen-icon lucide-japanese-yen\"\n     >\n      <g opacity={0.35}>\n       <path d=\"M12 9.5V21m0-11.5L6 3m6 6.5L18 3\" />\n       <path d=\"M6 15h12\" />\n       <path d=\"M6 11h12\" />\n      </g>\n\n      <m.path\n       d=\"M12 9.5V21m0-11.5L6 3m6 6.5L18 3\"\n       pathLength={1}\n       variants={vStroke}\n       initial=\"normal\"\n       animate={controls}\n      />\n      <m.path\n       d=\"M6 11h12\"\n       pathLength={1}\n       variants={midStroke}\n       initial=\"normal\"\n       animate={controls}\n      />\n      <m.path\n       d=\"M6 15h12\"\n       pathLength={1}\n       variants={baseStroke}\n       initial=\"normal\"\n       animate={controls}\n      />\n     </m.svg>\n    </m.div>\n   </LazyMotion>\n  );\n },\n);\n\nYenYuanIcon.displayName = \"YenYuanIcon\";\nexport { YenYuanIcon };\n",
-      "type": "registry:ui"
-    }
-  ]
-}
